recommender_project/requirements.txt
Django==5.0.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
numpy==1.25.0
